fix dump issue https://github.com/protocolbuffers/protobuf/issues/2954

+++ pkg/a2l/a2lhelpers.go
@@ -26,7 +26,7 @@ func a2lIntToIntType(integerValue parser.IIntegerValueContext) (result *IntType)
 
 	if tmpResult, err = strconv.ParseInt(rawString, int(base), 64); err == nil {
 		result = &IntType{
-			Value: tmpResult,
+			Value: int32(tmpResult),
 			Base:  base,
 			Size:  uint32(len(rawString)),
 		}
@@ -56,7 +56,7 @@ func a2lLongToLongType(integerValue parser.IIntegerValueContext) (result *LongTy
 
 	if tmpResult, err = strconv.ParseInt(rawString, int(base), 64); err == nil {
 		result = &LongType{
-			Value: tmpResult,
+			Value: int32(tmpResult),
 			Base:  base,
 			Size:  uint32(len(rawString)),
 		}
@@ -86,7 +86,7 @@ func numericToLongType(integerValue parser.INumericValueContext) (result *LongTy
 
 	if tmpResult, err = strconv.ParseInt(rawString, int(base), 64); err == nil {
 		result = &LongType{
-			Value: tmpResult,
+			Value: int32(tmpResult),
 			Base:  base,
 			Size:  uint32(len(rawString)),
 		}
@@ -148,7 +148,7 @@ func floatToFloatType(integerValue parser.INumericValueContext) (result *FloatTy
 	}
 
 	if tmpResult, err = strconv.ParseFloat(originalString, 64); err == nil {
-		result.Value = tmpResult
+		result.Value = float32(tmpResult)
 	} else {
 		panic(err)
 	}
