+++ pkg/a2l/a2mlhelpers.go
@@ -35,7 +35,7 @@ func arraySpecifierToLongType(integerValue parser.IArraySpecifierContext) (resul
 
 	if tmpResult, err = strconv.ParseInt(rawString, int(base), 64); err == nil {
 		result = &LongType{
-			Value: tmpResult,
+			Value: int32(tmpResult),
 			Base:  base,
 			Size:  uint32(len(rawString)),
 		}
